QMZJ/backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
eventlet==0.39.1
exceptiongroup==1.2.2
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
interchange==2021.0.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
monotonic==1.6
mysql-connector-python==9.2.0
neo4j==5.28.1
packaging==24.2
pansi==2024.11.0
pillow==11.1.0
py2neo==2021.2.4
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.8.0
PyMySQL==1.1.1
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.2
requests==2.32.3
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
wsproto==1.2.0
zhipuai==2.1.5.20250410
zope.event==5.0
zope.interface==7.2
